pyproject.toml
[tool.poetry]
name = "rpi-enviro-monitor"
version = "0.0.1"
description = "Python client for the Pimoroni Enviro for Raspberry Pi"
authors = ["Jordan Cook"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"

# Library dependencies
cffi = "^1.14"
fonts = "*"
font-roboto = "*"
loguru = "^0.5"
numpy = {version = ">=1.20", python = "<3.11"}
paho-mqtt = "^1.6"
pillow = "^8.0"
pyyaml = "^6.0"

# Hardware dependencies/drivers
ltr559 = "^0.1"
pimoroni-bme280 = "^0.1"
smbus = "^1.1"
sounddevice = "^0.4"
spidev = "^3.5"
st7735 = "^0.0.4"

[tool.poetry.dev-dependencies]
coverage = "^6.0"
pre-commit = "^2.12"
pytest = "^6.2"
pytest-cov = ">=2.11"

[tool.poetry.scripts]
enviro-run = 'rpi_enviro_monitor.console:run'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
skip-string-normalization = true
line-length = 100

[tool.isort]
profile = 'black'
line_length = 100
skip_gitignore = true

[tool.mypy]
ignore_missing_imports = true
files = ['rpi_enviro_monitor']
